requirements.txt
gunicorn==20.0.4
falcon==2.0
pyyaml==5.3.1
google-cloud-datastore==2.1.0
